[MODIFY] 클라우드 서버용 프록시 설정
