+++ package.json
@@ -46,5 +46,6 @@
     "prettier": "^3.5.3",
     "vite": "^6.2.0",
     "vite-plugin-pwa": "^1.0.0"
-  }
+  },
+  "proxy": "http://localhost:8080"
 }
