+++ src/components/User/OAuth2Success.jsx
@@ -7,7 +7,7 @@ const OAuth2Success = () => {
     useEffect(() => {
         const checkLogin = async () => {
             try {
-                const res = await fetch("http://localhost:8080/api/auth/check", {
+                const res = await fetch(`/api/auth/check`, {
                     credentials: "include",
                 });
